+++ java/src/main/java/com/google/api/services/samples/authorizedbuyers/marketplace/v1/buyers/proposals/PatchProposals.java
@@ -42,7 +42,7 @@ import net.sourceforge.argparse4j.inf.Namespace;
 public class PatchProposals {
 
   public static void execute(AuthorizedBuyersMarketplace marketplaceClient, Namespace parsedArgs) {
-    Integer accountId = parsedArgs.getInt("account_id");
+    Long accountId = parsedArgs.getLong("account_id");
     String proposalId = parsedArgs.getString(("proposal_id"));
     String name = String.format("buyers/%d/proposals/%s", accountId, proposalId);
     Long proposalRevision = parsedArgs.getLong("proposal_revision");
@@ -88,7 +88,7 @@ public class PatchProposals {
                 + "will be used to construct the name used as a path parameter for the "
                 + "proposals.patch request.")
         .required(true)
-        .type(Integer.class);
+        .type(Long.class);
     parser
         .addArgument("-p", "--proposal_id")
         .help(
